WDSUS-44 Move wpcli functions from class to separate files
